requirements.txt
networkx>=3.4
nx-parallel>=0.2
nx-cugraph-cu12
graphblas-algorithms>=2023.10.0
openai
python-dotenv
lxml
openpyxl
pandas
jupyterlab
nltk
dvc
tiktoken
rdflib
scikit-learn>=1.4.0 # there is a bug in 1.3 that causes f1 score to be wrong
pyarrow
ipytree
wandb
accelerate
bitsandbytes
transformers==4.50.3
chromadb
cohere==4.57
fasttext
pytest
sentencepiece
ete3[treeview]
mlxtend
dcor
spacy
numpy==1.26.*
bitsandbytes==0.45.1
datasets
tf-keras
sentence_transformers
parallelbar